refactor(llm): remove leftover compatibility surface
