+++ packages/llm/src/processor/sink-projection.ts
@@ -2,7 +2,7 @@ import { Operational, type Run, type Sink, type Tool } from "@openomni/protocol"
 import { Bus } from "@openomni/session";
 import { generateId } from "./contracts.js";
 
-export interface ProjectedSink extends Sink {
+interface ProjectedSink extends Sink {
   flush(): Promise<void>;
 }
