+++ packages/llm/src/provider/index.ts
@@ -189,13 +189,4 @@ export namespace Provider {
   }
 }
 
-export {
-  getSDK,
-  getLanguage,
-} from "./sdk";
-
-export { fetchProxyModels, enrichWithCatalog } from "./proxy-models";
-
 export { ModelsDev } from "../model";
-
-export { ProviderTransform } from "./transform";
